performance improvement but collision issue

+++ cmd/body/forcecalc.go
@@ -1,7 +1,7 @@
 package body
 
 type ForceCalcResult struct {
-	dist float32
+	dist float64
 	collided bool
 }
 
@@ -9,6 +9,6 @@ func NoCollision() ForceCalcResult {
 	return ForceCalcResult{0, false}
 }
 
-func Collision(dist float32) ForceCalcResult {
+func Collision(dist float64) ForceCalcResult {
 	return ForceCalcResult{dist, true}
 }
